Fix regenerate components
